transform: move spectral transforms and add from_types attr
